Now support sighash

+++ GetCoinbaseParts.go
@@ -83,11 +83,11 @@ func GetCoinbaseParts(height uint32, coinbaseValue uint64, defaultWitnessCommitm
 }
 
 func makeCoinbaseInputTransaction(coinbaseData []byte) []byte {
-	buf := make([]byte, 32)                              // 32 bytes - All bits are zero: Not a transaction hash reference
-	buf = append(buf, []byte{0xff, 0xff, 0xff, 0xff}...) // 4 bytes - All bits are ones: 0xFFFFFFFF
-	buf = append(buf, VarInt(len(coinbaseData))...)      // Length of the coinbase data, from 2 to 100 bytes
-	buf = append(buf, coinbaseData...)                   // Arbitrary data used for extra nonce and mining tags. In v2 blocks; must begin with block height
-	buf = append(buf, []byte{0xff, 0xff, 0xff, 0xff}...) //  4 bytes = Set to 0xFFFFFFFF
+	buf := make([]byte, 32)                                 // 32 bytes - All bits are zero: Not a transaction hash reference
+	buf = append(buf, []byte{0xff, 0xff, 0xff, 0xff}...)    // 4 bytes - All bits are ones: 0xFFFFFFFF
+	buf = append(buf, VarInt(uint64(len(coinbaseData)))...) // Length of the coinbase data, from 2 to 100 bytes
+	buf = append(buf, coinbaseData...)                      // Arbitrary data used for extra nonce and mining tags. In v2 blocks; must begin with block height
+	buf = append(buf, []byte{0xff, 0xff, 0xff, 0xff}...)    //  4 bytes = Set to 0xFFFFFFFF
 	return buf
 }
 
@@ -154,7 +154,7 @@ func makeCoinbaseOutputTransactions(coinbaseValue uint64, defaultWitnessCommitme
 
 	binary.LittleEndian.PutUint64(buf[0:], coinbaseValue)
 
-	buf = append(buf, VarInt(len(lockingScript))...)
+	buf = append(buf, VarInt(uint64(len(lockingScript)))...)
 	buf = append(buf, lockingScript...)
 
 	numberOfTransactions := 1
@@ -167,7 +167,7 @@ func makeCoinbaseOutputTransactions(coinbaseValue uint64, defaultWitnessCommitme
 			log.Printf("Error decoding witness commitment: %+v", err)
 			return nil, err
 		}
-		wcl := VarInt(len(wc))
+		wcl := VarInt(uint64(len(wc)))
 		buf = append(buf, wcl...)
 		buf = append(buf, wc...)
 	}
@@ -177,11 +177,11 @@ func makeCoinbaseOutputTransactions(coinbaseValue uint64, defaultWitnessCommitme
 
 		byteArr := make([]byte, 8) // 8 bytes of 0 = 0 satoshis.
 		buf = append(buf, byteArr...)
-		buf = append(buf, VarInt(len(minerIDBytes))...)
+		buf = append(buf, VarInt(uint64(len(minerIDBytes)))...)
 		buf = append(buf, minerIDBytes...)
 	}
 
-	buf = append(VarInt(numberOfTransactions), buf...)
+	buf = append(VarInt(uint64(numberOfTransactions)), buf...)
 	return buf, nil
 }
 
@@ -209,7 +209,7 @@ func makeCoinbase1(height uint32, coinbaseText string) []byte {
 	buf = append(buf, make([]byte, 32)...)               // Transaction hash - 4 bytes all bits are zero
 	buf = append(buf, []byte{0xff, 0xff, 0xff, 0xff}...) // Coinbase data size - 4 bytes - All bits are ones: 0xFFFFFFFF (ffffffff)
 
-	buf = append(buf, VarInt(len(arbitraryData)+spaceForExtraNonce)...) // Length of the coinbase data, from 2 to 100 bytes
+	buf = append(buf, VarInt(uint64(len(arbitraryData)+spaceForExtraNonce))...) // Length of the coinbase data, from 2 to 100 bytes
 	buf = append(buf, arbitraryData...)
 
 	return buf
